fix: storybookのディレクトリ構成を変更した
