+++ app/src/components/block/Block.tsx
@@ -3,13 +3,13 @@ import Paragraph from '@/components/block/Paragraph'
 import { Block as BlockData } from '@/type/page/block/block'
 import { RichText as RichTextData } from '@/type/page/block/richText'
 
-import BookMark from './BookMark'
+import BookMark from './bookMark/BookMark'
 import BulletedList from './bulletedList/BulletedList'
 import BulletedListItem from './bulletedList/BulletedListItem'
 import Callout from './callout/Callout'
 import Code from './code/Code'
 import Divider from './divider/Divider'
-import Equation from './Equation'
+import Equation from './equation/Equation'
 import Image from './image/Image'
 import NumberedList from './numberedList/NumberedList'
 import NumberedListItem from './numberedList/NumberedListItem'
