+++ app/src/components/block/Header.stories.tsx
@@ -1,12 +1,11 @@
-import Anchor from '../components/common/Anchor'
+import Header from '../layout/Header'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/Anchor',
-    component: Anchor,
+    title: 'Example/Header',
+    component: Header,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -15,15 +14,20 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof Anchor>
+} satisfies Meta<typeof Header>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
+// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
 export const Default: Story = {
-    args: {
-        children: 'Anchor',
-        href: 'https://www.google.com',
-        isInnerLink: false,
+    args: {},
+    render() {
+        return (
+            <div>
+                <meta.component></meta.component>
+                <main>{new Array(1000).fill(0).map((v, i) => `${i} `)}</main>
+            </div>
+        )
     },
 }
