+++ app/src/components/block/bookMark/BookMark.stories.tsx
@@ -1,12 +1,11 @@
-import BookMark from '../components/block/BookMark'
+import Equation from '../equation/Equation'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/BookMark',
-    component: BookMark,
+    title: 'Example/Equation',
+    component: Equation,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -15,13 +14,13 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof BookMark>
+} satisfies Meta<typeof Equation>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
     args: {
-        href: 'https://www.google.com',
+        math: 'x^2 + y^2 = z^2',
     },
 }
