+++ app/src/components/block/equation/Equation.stories.tsx
@@ -1,12 +1,11 @@
-import TagList from '../components/common/tag/TagList'
+import BookMark from '../bookMark/BookMark'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/TagList',
-    component: TagList,
+    title: 'Example/BookMark',
+    component: BookMark,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -15,26 +14,13 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof TagList>
+} satisfies Meta<typeof BookMark>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
     args: {
-        tagData: [
-            {
-                name: 'Tag 1',
-                color: 'red',
-            },
-            {
-                name: 'Tag 2',
-                color: 'blue',
-            },
-            {
-                name: 'Tag 3',
-                color: 'green',
-            },
-        ],
+        href: 'https://www.google.com',
     },
 }
