+++ app/src/components/block/equation/Equation.tsx
