+++ app/src/components/common/anchor/Anchor.stories.tsx
@@ -1,15 +1,11 @@
-
-import { useState } from 'react'
-
-import Card from '../components/common/Card'
-import Modal from '../components/common/Modal'
+import Anchor from './Anchor'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/Modal',
-    component: Modal,
+    title: 'Example/Anchor',
+    component: Anchor,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -18,26 +14,15 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof Modal>
+} satisfies Meta<typeof Anchor>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
     args: {
-        children: 'Hello Modal',
-        show: true,
-        setShow: () => {},
-    },
-    render() {
-        const [show, setShow] = useState(true)
-        return (
-            <>
-                <div>{'hoge '.repeat(10000)}</div>
-                <meta.component show={show} setShow={setShow}>
-                    <Card></Card>
-                </meta.component>
-            </>
-        )
+        children: 'Anchor',
+        href: 'https://www.google.com',
+        isInnerLink: false,
     },
 }
