+++ app/src/components/common/button/Button.stories.tsx
@@ -1,12 +1,11 @@
-import ImageBase from '../components/common/ImageBase'
+import { Button } from './Button'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/ImageBase',
-    component: ImageBase,
+    title: 'Example/Button',
+    component: Button,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -14,15 +13,33 @@ const meta = {
     // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-    argTypes: {},
-} satisfies Meta<typeof ImageBase>
+    argTypes: {
+        mode: {
+            control: 'select',
+            options: ['primary', 'secondary', 'orange'],
+        },
+    },
+} satisfies Meta<typeof Button>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
-export const Default: Story = {
+// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
+export const Primary: Story = {
     args: {
-        src: 'https://upload.wikimedia.org/wikipedia/commons/9/92/Ja-fukushima-ohuchijuku-7.jpg',
-        alt: 'ImageBase alt text',
+        mode: 'primary',
+        children: 'Button',
+    },
+}
+export const Secondary: Story = {
+    args: {
+        mode: 'secondary',
+        children: 'Button',
+    },
+}
+export const Orange: Story = {
+    args: {
+        mode: 'orange',
+        children: 'Button',
     },
 }
