+++ app/src/components/common/card/Card.stories.tsx
@@ -1,12 +1,11 @@
-import Tag from '../components/common/tag/Tag'
+import Card from './Card'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/Tag',
-    component: Tag,
+    title: 'Example/Card',
+    component: Card,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -15,25 +14,19 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof Tag>
+} satisfies Meta<typeof Card>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-    args: {
-        name: 'タグ',
-        color: 'blue',
-    },
-}
-
-export const WithLink: Story = {
-    args: { name: 'タグ', color: 'blue' },
+    args: {},
     render() {
         return (
-            <a href='#' className='hover:underline'>
-                <meta.component name='タグ' color='blue'></meta.component>
-            </a>
+            <meta.component>
+                <h3>hogehoge</h3>
+                <p>fugafuga fugafuga</p>
+            </meta.component>
         )
     },
 }
