+++ app/src/components/common/carousel/Carousel.tsx
@@ -7,7 +7,7 @@ import { FaAngleLeft, FaAngleRight } from 'react-icons/fa'
 import { Tag as TagType } from '@/type/page/page'
 
 import dateToStr from '../../../util/dateToStr'
-import Image from '../ImageBase'
+import ImageBase from '../imageBase/ImageBase'
 import TagList from '../tag/TagList'
 import './Carousel.style.scss'
 
@@ -128,12 +128,12 @@ const Carousel = ({ data, ...props }: CarouselProps) => {
                             key={index}
                             className='relative min-w-full'
                         >
-                            <Image
+                            <ImageBase
                                 src={d.image}
                                 alt={`image of ${d.title}`}
                                 className='aspect-video h-full snap-center'
                                 objectFit='cover'
-                            ></Image>
+                            ></ImageBase>
                             <div className='absolute bottom-0 flex w-full flex-col gap-2 bg-black/25 p-3 text-sm text-white md:px-4'>
                                 <p>{d.title}</p>
                                 <div className='flex items-end justify-between gap-4'>
