+++ app/src/components/common/imageBase/ImageBase.style.scss
