+++ app/src/components/common/tag/TagList.stories.tsx
@@ -1,12 +1,11 @@
-import Card from '../components/common/Card'
+import TagList from './TagList'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    title: 'Example/Card',
-    component: Card,
+    title: 'Example/TagList',
+    component: TagList,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -15,19 +14,26 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof Card>
+} satisfies Meta<typeof TagList>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-    args: {},
-    render() {
-        return (
-            <meta.component>
-                <h3>hogehoge</h3>
-                <p>fugafuga fugafuga</p>
-            </meta.component>
-        )
+    args: {
+        tagData: [
+            {
+                name: 'Tag 1',
+                color: 'red',
+            },
+            {
+                name: 'Tag 2',
+                color: 'blue',
+            },
+            {
+                name: 'Tag 3',
+                color: 'green',
+            },
+        ],
     },
 }
